Match Property filter added
The user can specify which properties to analyse.

the template secition that retrieved the label for the match property was broken,

so the function: map_match_property_to_label, gives them human readable names.

I also added handling for empty methodsubtype values in the query response, as this was causing issues.